requirements.txt
matplotlib
wandb
albumentations
lightning
python-dotenv
numpy
prettytable
codecov
flake8
ftfy
interrogate
isort==4.3.21
pytest
regex
xdoctest>=0.10.0
yapf
